requirements-dev.txt
boto3>=1.34.0
botocore>=1.34.0
moto[s3]>=5.0.0
pytest>=7.0.0
pytest-mock>=3.0.0
python-dotenv